tidy tests
